Set  "additionalProperties": false where it makes sense
When additionalProperties is not set, the user will not get warnings for typos.
or subfields of typos.

This might potentially break something if schemas are used for validation
and invalid schemas are in use. It might also make it somewhat harder for
local testing of new properties, if strict validation is required for some
operation. I have not seen any instances where strict validation causes
failiures, but I have not tried everything.